requirements.txt
python-dotenv==1.0.0
discord.py==2.3.2
better_profanity==0.7.0
numpy==1.26.0
validators==0.22.0
coverage==7.3.2
pytest==7.4.2
pylint==3.0.1
psutil==5.9.5
apscheduler==3.10.4
pyshorteners~=1.0.1
quickchart.io==1.0.0
bard==0.1
bardapi==0.1.38
